Move {template}.js to templates folder, upd main.js
