+++ js/main.js
@@ -1,10 +1,10 @@
-import intro from './intro';
-import greeting from './greeting';
-import rules from './rules';
-import game1 from './game-1';
-import game2 from './game-2';
-import game3 from './game-3';
-import stats from './stats';
+import intro from './templates/intro';
+import greeting from './templates/greeting';
+import rules from './templates/rules';
+import game1 from './templates/game-1';
+import game2 from './templates/game-2';
+import game3 from './templates/game-3';
+import stats from './templates/stats';
 
 (() => {
