+++ js/templates/greeting.js
