+++ js/templates/stats.js
